좌석예약 선택창 필요값 불러오기

+++ src/main/java/com/kh/skycastle/coupon/model/vo/Coupon.java
@@ -20,6 +20,6 @@ public class Coupon {
 	private String couponName;
 	private Date startDate;
 	private int expirationPeriod;
-	private int discountRate;
+	private double discountRate;
 
 }
